Confirm deletion of Azure Cluster.

+++ src/ClusterBootstrap/az_tools.py
@@ -357,7 +357,9 @@ def run_command( args, command, nargs, parser ):
         create_cluster()
 
     elif command == "delete":
-        delete_group()
+        response = raw_input ("!!! WARNING !!! You are performing a dangerous operation that will permanently delete the entire Azure DL Workspace cluster. Please type (DELETE) in ALL CAPITALS to confirm the operation ---> ")
+		if response == "DELETE":
+			delete_group()
 
     elif command == "genconfig":
         gen_cluster_config("cluster.yaml")
